Edited lights and added very simple bot for testing
Lights now have a light span which is between 0 and 360 and a direction.
Small changes to light shaders so that lights now light only the radius
they are given. Edited AABB class for some small optimizations. One
random bot is added for testing AI controls. In order to be proper
though an AI thread needs to be created.