Added Geometry class, some small changes and some code refactoring
Added Geometry class which holds all the points and lines of an object.
All rotations, scales and transformations are done trought it. Added
normal vector in the Line class so that it is precalculated. Renamed
ObjectManager to scene class, because later I will add a object manager
class which will deal with memory allocation. Minor change in Shader
class so that it is easier to load shaders that way.